+++ CHANGELOG.md
@@ -21,3 +21,6 @@ New technique of working but the same way to code.
 Adding feature: Ability to get the data stored in the provider just before the last update.
 Adding feature: Adding [statePoint]s to the [AmRefreshWidget] and ability to get the [AmRefreshWidgetController].
 
+## 2.0.1
+Adding feature: Adding [AmMemory] extention class that provides data access;
+    Save or Get Data from everywhere using [ID];